+++ backend/pcloud-domain/src/test/java/com/domain/common/PriceTest.java
@@ -1,4 +1,4 @@
-package com.domain.domains.common;
+package com.domain.common;
 
 import com.domain.common.Price;
 import com.domain.popups.exception.PopupsException;
