+++ backend/pcloud-domain/src/test/java/com/domain/infrastructure/CustomTagJpaRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.domain.domains.customtag.infrastructure;
+package com.domain.infrastructure;
 
 import com.domain.customtag.domain.CustomTag;
 import com.domain.customtag.infrastructure.CustomTagJpaRepository;
