+++ backend/pcloud-domain/src/test/java/com/domain/popups/domain/vo/LatitudeTest.java
@@ -1,6 +1,6 @@
-package com.domain.domains.popups.domain.vo;
+package com.domain.popups.domain.vo;
 
-import com.domain.popups.domain.vo.Longitude;
+import com.domain.popups.domain.vo.Latitude;
 import com.domain.popups.exception.PopupsException;
 import com.domain.popups.exception.PopupsExceptionType;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -16,26 +16,26 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
-class LongitudeTest {
+class LatitudeTest {
 
     @Test
-    void 경도를_생성한다() {
+    void 위도를_생성한다() {
         // given
-        BigDecimal value = BigDecimal.valueOf(130);
+        BigDecimal value = BigDecimal.valueOf(35);
 
         // when
-        Longitude result = Longitude.from(value);
+        Latitude result = Latitude.from(value);
 
         // then
         assertThat(result.getValue()).isEqualTo(value);
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {"124", "133"})
-    void 한국_범위가_벗어난_경도는_예외를_발생시킨다(final String value) {
+    @ValueSource(strings = {"32", "40"})
+    void 한국_범위가_벗어난_위도는_예외를_발생시킨다(final String value) {
         // when & then
-        assertThatThrownBy(() -> Longitude.from(value))
+        assertThatThrownBy(() -> Latitude.from(value))
                 .isInstanceOf(PopupsException.class)
-                .hasMessageContaining(PopupsExceptionType.INVALID_LONGITUDE.message());
+                .hasMessageContaining(PopupsExceptionType.INVALID_LATITUDE.message());
     }
 }
